works

+++ easy_tg_bot/put_file_conv.py
@@ -12,9 +12,9 @@ from .utils.utils import get_keyboard, end_conversation
 from .send import send_keyboard
 
 from .file_handler import FileHandler
-from .admin import is_user_admin, admin
-# entry points for the conversations
-from .admin import PUT_TEXT_BUTTON
+from .roles import check_role, DEFAULT_ADMIN_ROLES
+# + entry point for the conversations
+from .admin import admin, ADD_TEXT_BUTTON
 
 # file handlers
 from .text_handler import text_handler
@@ -37,8 +37,8 @@ class PutFileConversation:
     ):
         self.file_handler = file_handler
         self.entry_point_query_pattern = entry_point_query_pattern
-        self.restrict_function = is_user_admin
         self.conversation_done_callback_function = admin
+        self.restrict_function = lambda u, c: check_role(c, DEFAULT_ADMIN_ROLES)
         self.INPUT_FILE = 0
 
         self.put_file_message = f"put_{file_handler.prefix}_file_message"  # text.xlsx
@@ -65,12 +65,15 @@ class PutFileConversation:
         )
 
     async def end(self, update: Update, context: CallbackContext):
+        if not self.restrict_function(update, context):
+            return ConversationHandler.END  # TODO decorators
+
         await self.conversation_done_callback_function(update, context)
         return ConversationHandler.END
 
     async def start_conversation(self, update: Update, context: CallbackContext):
         if not self.restrict_function(update, context):
-            return ConversationHandler.END  # TODO decorators
+            return ConversationHandler.END
 
         query = update.callback_query
         await query.answer()
@@ -86,6 +89,9 @@ class PutFileConversation:
     async def file_error_response(
         self, update: Update, context: CallbackContext, file_path, message_text_id
     ):
+        if not self.restrict_function(update, context):
+            return ConversationHandler.END 
+
         await self.file_handler.delete_file(file_path)
         keyboard = get_keyboard(
             context,
@@ -96,7 +102,7 @@ class PutFileConversation:
 
     async def receive_file(self, update: Update, context: CallbackContext):
         if not self.restrict_function(update, context):
-            return
+            return ConversationHandler.END
 
         try:
             file_path = await self.file_handler.download_file(update)
@@ -132,5 +138,5 @@ class PutFileConversation:
 
 
 # My handlers
-put_text_file_conv = PutFileConversation(text_handler, PUT_TEXT_BUTTON)
+put_text_file_conv = PutFileConversation(text_handler, ADD_TEXT_BUTTON)
 put_text_file_conv.register_handler()
